% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shinyga_auth.R
\name{appURL}
\alias{appURL}
\title{Retrieve Shiny URL}
\usage{
appURL(session)
}
\arguments{
\item{session}{The session object}
}
\description{
Get token if it's previously stored, else prompt user to get one.
}
\keyword{internal}

