% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shinyga_macros.R
\name{authDropdownRow}
\alias{authDropdownRow}
\title{authDropdownRow - creates a row of GA authentication menus}
\usage{
authDropdownRow(account.id = "accounts", web.prop.id = "web.prop",
  view.id = "view")
}
\arguments{
\item{account.id}{The shiny id for accounts. Then available at input$<account.id>.}

\item{web.prop.id}{The shiny id for web properties. Then available at input$<web.prop.id>.}

\item{view.id}{The shiny id for views. Then available at input$<view.id>.}
}
\value{
A shinydashboard function that generates necessary HTML.
}
\description{
Use with renderAuthDropdownRow()
Creates a row of select boxes for GA Account, Webproperty and View.
id needed for GA fetches is then available in input${view.id}
}
\examples{
\dontrun{
 ## server.r
shinyServer(function(input, output, session) {

    renderAuthDropdownRow(ga.table = ShinyMakeGAProfileTable(),
                          input = input,
                          session = session)

}

 ## ui.r

 library(shinydashboard)
 dashboardBody(authDropdownRow())

 }
}
\seealso{
Other shiny macro functions: \code{\link{renderAuthDropdownRow}}
}

