% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shinyga_auth.R
\name{createCode}
\alias{createCode}
\title{Creates a random character code}
\usage{
createCode(seed = NULL, num = 20)
}
\arguments{
\item{seed}{The seed for random code.}

\item{num}{Number of characters in code.}
}
\value{
A random string of digits and characters.
}
\description{
Creates a random character code
}
\examples{
\dontrun{
securityCode <- createCode()
shinyServer(function(input, output, session)){

  AuthCode <- reactive({

      authReturnCode(session, securityCode)

  })

  output$AuthGAURL <- renderUI({

       a("Click Here to Authorise Your Google Analytics Access",
          href=shinygaGetTokenURL(securityCode)
          )
       })

  AccessToken <- reactive({
      validate(
        need(AuthCode(), "Authenticate To See")
      )

      access_token <- shinygaGetToken(code = AuthCode())

      token <- access_token$access_token

    })
}
}
}
\seealso{
Other authentication functions: \code{\link{authReturnCode}};
  \code{\link{shinygaGetTokenURL}};
  \code{\link{shinygaGetToken}}
}

