% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyAnalytics.R
\name{enableTracking}
\alias{enableTracking}
\title{Enable GA tracking}
\usage{
enableTracking(trackingID)
}
\arguments{
\item{trackingID}{Unique website tracking ID given by Google analytics}
}
\description{
This function must be called at the begining of the app's server code. It calls
\code{ga(create, trackingID, auto)}
}

