% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyAnalytics.R
\name{gaTrack}
\alias{gaTrack}
\title{Input tracking}
\usage{
gaTrack(inputId, session, category, action, label = inputId)
}
\arguments{
\item{inputId}{Input unique ID string}

\item{category}{Event category}

\item{action}{Event action}

\item{label}{Event label}

\item{input}{Shiny server reactive input value}
}
\description{
Sends an event with given category, action and label whenever
the given input changes. This function must be called in the app's server side.
}

