% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shinyga_fetchdata.R
\name{getAndMergeGAAccounts}
\alias{getAndMergeGAAccounts}
\title{getAndMergeGAAccounts - creates dataframe of all GA accounts/webproperties/views}
\usage{
getAndMergeGAAccounts(token)
}
\arguments{
\item{token}{Token passed from shinygaGetToken().}
}
\value{
A dataframe of GA Account, Webproperty and View ID.
}
\description{
creates dataframe of GA accounts, webprops, views that is passed to rollupGA
}
\examples{
Views <- getAndMergeGAAccounts(123456)
}
\seealso{
Other fetch data functions: \code{\link{MEgetData}};
  \code{\link{processManagementData}};
  \code{\link{rollupGA}}; \code{\link{shinygaGetAccounts}};
  \code{\link{shinygaGetAdWords}};
  \code{\link{shinygaGetCustomDataSources}};
  \code{\link{shinygaGetCustomDimensions}};
  \code{\link{shinygaGetCustomMetrics}};
  \code{\link{shinygaGetFilters}};
  \code{\link{shinygaGetGoals}};
  \code{\link{shinygaGetProfiles}};
  \code{\link{shinygaGetSegments}};
  \code{\link{shinygaGetUsers}};
  \code{\link{shinygaGetWebProperties}}
}

