% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shinyga_messages.R
\name{initMessageData}
\alias{initMessageData}
\title{initMessageData - creates Message object}
\usage{
initMessageData(text = c("Welcome! Authenticate to get started."),
  icon = c("smile-o"), status = c("info"))
}
\arguments{
\item{text}{Text of message.}

\item{icon}{Icon taken from shinydashboard.}

\item{status}{Color of message. See Details.}
}
\value{
messageData Object to be passed to addMessageData().
}
\description{
Use this first before addMessageData adds to the notification menu when stuff happens
}
\examples{
\dontrun{

## server.r
shinyServer(function(input, output, session) {

    messageData <- initMessageData()

    addMessageData(messageData, "A new message!")

      output$messageMenu <- renderDropdownMenu({

      message_df <- data.frame(text = messageData$text,
                               icon = messageData$icon,
                               status = messageData$status)

     msgs <- apply(message_df, 1, function(row){
                   notificationItem(text = row[['text']],
                                    icon = icon(row[['icon']]),
                                    status = row[['status']])

                   })

     dropdownMenu(type="notifications", .list = msgs)

     })

}

## ui.r

dashboardHeader(title = "GA Forecast",
                dropdownMenuOutput("messageMenu"))
}
}
\seealso{
\url{http://rstudio.github.io/shinydashboard/appearance.html#icons}

Other message functions: \code{\link{addMessageData}}
}

