% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shinyga_fetchdata.R
\name{MEgetData}
\alias{MEgetData}
\title{MEgetData - does the actual fetch of GA data, called from rollupGA}
\usage{
MEgetData(ids, start.date = format(Sys.time(), "\%Y-\%m-\%d"),
  end.date = format(Sys.time(), "\%Y-\%m-\%d"), metrics = "ga:visits",
  dimensions = "ga:date", sort = "", filters = "", segment = "",
  fields = "", start = 1, max, date.format = "\%Y-\%m-\%d",
  messages = TRUE, batch, walk = FALSE, output.raw, output.formats,
  return.url = FALSE, rbr = FALSE, envir = .GlobalEnv, token)
}
\arguments{
\item{ids}{The view ID}

\item{start.date}{Start date of data fetch}

\item{end.date}{End date of data fetch}

\item{metrics}{Metrics called, in GA API format (e.g. ga:visits,ga:pageviews)}

\item{dimensions}{Dimensions called, in GA API format (e.g. ga:date, ga:source)}

\item{sort}{Sort the results before returning data}

\item{filters}{Filters of data}

\item{segment}{Any segments}

\item{fields}{What fields from the API should be returned}

\item{start}{What start index of the data, used for walking through big results}

\item{max}{Maximum number of results to fetch}

\item{date.format}{Date format of results}

\item{messages}{Feedback messages}

\item{batch}{If results over 10000 limit, whether to batch them}

\item{walk}{If results are sampled, whether to do daily fetches to avoid sampling}

\item{output.raw}{Whether the results should be outputed not parsed into data.frame}

\item{output.formats}{Formats of data output}

\item{return.url}{Whether to return a URL}

\item{rbr}{Row by Row setting, will return NAs for empty rows}

\item{envir}{Which environment the data will be}

\item{The}{token needed to fetch the data}
}
\value{
A dataframe of GA data.
}
\description{
An internal function called from rollupGA, this does the actual fetch from GA.
}
\examples{
gadata <- MEgetData(123456, token=gatoken)
}
\seealso{
Refer to the dimensions and metric Google help file for more details.
\url{https://developers.google.com/analytics/devguides/reporting/core/dimsmets}

Other fetch data functions: \code{\link{getAndMergeGAAccounts}};
  \code{\link{processManagementData}};
  \code{\link{rollupGA}}; \code{\link{shinygaGetAccounts}};
  \code{\link{shinygaGetAdWords}};
  \code{\link{shinygaGetFilters}};
  \code{\link{shinygaGetGoals}};
  \code{\link{shinygaGetProfiles}};
  \code{\link{shinygaGetSegments}};
  \code{\link{shinygaGetWebProperties}}
}

