% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shinyga_macros.R
\name{metricSelect}
\alias{metricSelect}
\title{metricSelect - creates a selection of which metric}
\usage{
metricSelect(inputId = "metric_choice")
}
\arguments{
\item{inputId}{The id of the input.  Then available in input$<inputId>.}
}
\value{
A shinydashboard function that generates necessary HTML.
}
\description{
Creates a select box with GA metrics. Metric var then available in input$<inputId>
}
\examples{
\dontrun{
 ## ui.r

 library(shinydashboard)
 dashboardBody(authDropdownRow(),
               metricSelect())
         }
}

