% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shinyga_fetchdata.R
\name{processManagementData}
\alias{processManagementData}
\title{Take GA API output and parses it into data.frame}
\usage{
processManagementData(url, keep)
}
\arguments{
\item{url}{The GA API response URL.}

\item{keep}{Which colums to keep.}
}
\value{
A dataframe of GA data.
}
\description{
Internal function for returning a dataframe from API response
}
\examples{
\dontrun{
shinygaGetSegments = function(token, start=1, max=1000) {
url <- paste('https://www.googleapis.com/analytics/v3/management/segments',
             '?access_token=', token,
             '&start-index=', start,
             '&max-results=', max,
             sep='', collapse='')

             return(processManagementData(url,
                    c('id', 'segmentId', 'name', 'definition', 'created', 'updated')))
                    }
}
}
\seealso{
Other fetch data functions: \code{\link{MEgetData}};
  \code{\link{getAndMergeGAAccounts}};
  \code{\link{rollupGA}}; \code{\link{shinygaGetAccounts}};
  \code{\link{shinygaGetGoals}};
  \code{\link{shinygaGetProfiles}};
  \code{\link{shinygaGetSegments}};
  \code{\link{shinygaGetWebProperties}}
}

