% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shinyga_macros.R
\name{renderAuthDropdownRow}
\alias{renderAuthDropdownRow}
\title{renderAuthDropdownRow - creates a row of GA authentication menus}
\usage{
renderAuthDropdownRow(ga.table, input, session, account.id = "accounts",
  web.prop.id = "web.prop", view.id = "view")
}
\arguments{
\item{ga.table}{A table of GA profiles from getAndMergeGAAccounts().}

\item{input}{The shiny input object.}

\item{session}{The shiny session object.}

\item{account.id}{The shiny id for accounts. Then available at input$<account.id>.}

\item{web.prop.id}{The shiny id for web properties. Then available at input$<web.prop.id>.}

\item{view.id}{The shiny id for views. Then available at input$<view.id>.}
}
\value{
Nothing.
}
\description{
Use with authDropdownRow
Creates a row of select boxes for GA Account, Webproperty and View.
id needed for GA fetches is then available in input${view.id}
}
\examples{
\dontrun{
 ## server.r
shinyServer(function(input, output, session) {

    renderAuthDropdownRow(ga.table = ShinyMakeGAProfileTable(),
                          input = input,
                          session = session)

}

 ## ui.r

 library(shinydashboard)
 dashboardBody(authDropdownRow(),
               metricSelect())

 }
}
\seealso{
Shortcut using \code{\link{doAuthMacro}}.

Other shiny macro functions: \code{\link{authDropdownRow}};
  \code{\link{doSegmentMacro}}; \code{\link{metricSelect}}
}

