% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shinyga_auth.R
\name{renderAuthenticationMacro}
\alias{renderAuthenticationMacro}
\title{Shortcut all authentication functions}
\usage{
renderAuthenticationMacro(input, output, session, securityCode, client.id,
  client.secret, client.url, messageData)
}
\arguments{
\item{input}{The Shiny input object.}

\item{output}{The Shiny output object.}

\item{session}{The Shiny session object.}

\item{securityCode}{A security code generated by createCode}

\item{client.id}{From the Google API console}

\item{client.secret}{From the Google API console}

\item{client.url}{The URL of your application}

\item{messageData}{Object created from initMessageData}
}
\value{
Recreates all authentication functions needed for shinyServer call.
 Use in ui.r:
 output$AuthGAURL - URL users click on (uiOutput("AuthGAURL"))
 output$GAProfile - DataTable of GA Profie data (dataTableOutput("GAProfile"))
 Use in server.r:
 input$accounts - The accountId
 input$web.prop - The webprop Id
 input$view - The View Id (used in GA data fetches)
}
\description{
This takes all the authentication functions and renders them in a macro
  used for quickstart of dashboards. Trade flexibility for speed.
}
\examples{
\dontrun{
securityCode <- createCode()
CLIENT_ID <- "xxxxxxxxxxxx.apps.googleusercontent.com"
CLIENT_SECRET <- "xxxxxxxxxxxxxxx"
CLIENT_URL <- 'https://mark.shinyapps.io/ga-effect/'
CLIENT_URL <- 'http://127.0.0.1:6423' #comment out for local testing

shinyServer(function(input, output, session)){

  messageData <- initMessageData()

  renderAuthenticationMacro(input,
                            output,
                            session,
                            securityCode,
                            CLIENT_ID,
                            CLIENT_SECRET,
                            CLIENT_URL,
                            messageData)
  }
}
}
\seealso{
\code{\link{renderAuthDropdownRow}}
         \code{\link{initMessageData}}

Other authentication functions: \code{\link{authReturnCode}};
  \code{\link{createCode}};
  \code{\link{shinygaGetTokenURL}};
  \code{\link{shinygaGetToken}}
}

