% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shinyga.R
\docType{package}
\name{shinyga}
\alias{shinyga}
\alias{shinyga-package}
\title{shinyga: Easier Google Analytics Dashboards in Shiny.}
\description{
This package adds functions to be used within a Shiny project, that quickstart
importing and displaying Google API data.
}
\details{
Package focused on quick setup of Google Authentication APIs,
with initial focus on Google Analytics.
}
\section{Authentication functions}{

\code{\link{doAuthMacro}}
\code{\link{renderAuthDropdownRow}}
\code{\link{createCode}}
}

\section{Message Functions}{

\code{\link{initMessageData}}
\code{\link{addMessageData}}
}

\section{Fetch Data Functions}{

\code{\link{rollupGA}}
}

\section{Macro Functions}{

\code{\link{metricSelect}}
}

