% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shinyga_fetchdata.R
\name{shinygaGetAccounts}
\alias{shinygaGetAccounts}
\title{Get GA Account data}
\usage{
shinygaGetAccounts(token, start = 1, max = 1000)
}
\arguments{
\item{token}{Token passed from shinygaGetToken().}

\item{start}{Which data point to start from.}

\item{max}{Maximum number to fetch.}
}
\value{
If token exists, a dataframe of GA accounts.
}
\description{
Gets the accounts available from the user token
}
\examples{
\dontrun{
Accounts <- shinygaGetAccounts(123456)
}
}
\seealso{
Other fetch data functions: \code{\link{MEgetData}};
  \code{\link{getAndMergeGAAccounts}};
  \code{\link{processManagementData}};
  \code{\link{rollupGA}}; \code{\link{shinygaGetFilters}};
  \code{\link{shinygaGetGoals}};
  \code{\link{shinygaGetProfiles}};
  \code{\link{shinygaGetSegments}};
  \code{\link{shinygaGetWebProperties}}
}

