% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shinyga_fetchdata.R
\name{shinygaGetCustomDimensions}
\alias{shinygaGetCustomDimensions}
\title{Get GA customDimensions}
\usage{
shinygaGetCustomDimensions(token, accountId, webPropertyId, start = 1,
  max = 1000)
}
\arguments{
\item{token}{Token passed from shinygaGetToken()}

\item{accountId}{AccountId of webproperty}

\item{webPropertyId}{webPropertyId}

\item{max}{Maximum number to fetch}
}
\value{
If token exists, a dataframe of filters for the account.
}
\description{
Gets the Custom Dimensions available for each web property.
}
\examples{
\dontrun{
dims <- shinygaGetCustomDimensions(123456)
}
}
\seealso{
Other fetch data functions: \code{\link{MEgetData}};
  \code{\link{getAndMergeGAAccounts}};
  \code{\link{processManagementData}};
  \code{\link{rollupGA}}; \code{\link{shinygaGetAccounts}};
  \code{\link{shinygaGetAdWords}};
  \code{\link{shinygaGetCustomDataSources}};
  \code{\link{shinygaGetCustomMetrics}};
  \code{\link{shinygaGetFilters}};
  \code{\link{shinygaGetGoals}};
  \code{\link{shinygaGetProfiles}};
  \code{\link{shinygaGetSegments}};
  \code{\link{shinygaGetUsers}};
  \code{\link{shinygaGetWebProperties}}
}

