% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shinyga_fetchdata.R
\name{shinygaGetGoals}
\alias{shinygaGetGoals}
\title{Get GA Goals}
\usage{
shinygaGetGoals(token, accountId = "~all", webPropertyId = "~all",
  profileId = "~all", start = 1, max = 1000)
}
\arguments{
\item{token}{Token passed from shinygaGetToken()}

\item{accountId}{AccountId of webproperty, or all of them with ~all}

\item{webPropertyId}{webPropertyId, or all of them with ~all}

\item{profileId}{profileId (ViewId), or all of them with ~all}

\item{start}{which data point to start from}

\item{max}{Maximum number to fetch}
}
\value{
If token exists, a dataframe of GA Goals.
}
\description{
Gets the Goals available from the user token.
}
\examples{
\dontrun{
Views <- shinygaGetGoals(123456)
}
}
\seealso{
Other fetch data functions: \code{\link{MEgetData}};
  \code{\link{getAndMergeGAAccounts}};
  \code{\link{processManagementData}};
  \code{\link{rollupGA}}; \code{\link{shinygaGetAccounts}};
  \code{\link{shinygaGetFilters}};
  \code{\link{shinygaGetProfiles}};
  \code{\link{shinygaGetSegments}};
  \code{\link{shinygaGetWebProperties}}
}

