% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shinyga_fetchdata.R
\name{shinygaGetSegments}
\alias{shinygaGetSegments}
\title{Get GA Segments}
\usage{
shinygaGetSegments(token, start = 1, max = 1000)
}
\arguments{
\item{token}{Token passed from shinygaGetToken()}

\item{start}{which data point to start from}

\item{max}{Maximum number to fetch}
}
\value{
If token exists, a dataframe of GA segments.
}
\description{
Gets the Segments available from the user token.
}
\examples{
\dontrun{
Views <- shinygaGetSegments(123456)
}
}
\seealso{
Other fetch data functions: \code{\link{MEgetData}};
  \code{\link{getAndMergeGAAccounts}};
  \code{\link{processManagementData}};
  \code{\link{rollupGA}}; \code{\link{shinygaGetAccounts}};
  \code{\link{shinygaGetGoals}};
  \code{\link{shinygaGetProfiles}};
  \code{\link{shinygaGetWebProperties}}
}

