% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shinyga_auth.R
\name{shinygaGetTokenURL}
\alias{shinygaGetTokenURL}
\title{Returns the authentication URL}
\usage{
shinygaGetTokenURL(state, client.id = CLIENT_ID,
  client.secret = CLIENT_SECRET, redirect.uri = CLIENT_URL,
  scope = c("https://www.googleapis.com/auth/analytics",
  "https://www.googleapis.com/auth/analytics.readonly"))
}
\arguments{
\item{state}{A string you pass to check authentication is correct.}

\item{client.id}{The client ID taken from the Google API Console.}

\item{client.secret}{The client secret taken from the Google API Console.}

\item{redirect.uri}{The URL of where your Shiny application sits, that will read state parameter.}

\item{scope}{Character vector of URLs for the Google API you want to activate.
Defaults to \code{c("https://www.googleapis.com/auth/analytics",
              "https://www.googleapis.com/auth/analytics.readonly")}}
}
\value{
The URL a user should click on to start authentication.
}
\description{
Returns the authentication URL
}
\examples{
\dontrun{
securityCode <- createCode()
shinyServer(function(input, output, session)){

  AuthCode <- reactive({

      authReturnCode(session, securityCode)

  })

  output$AuthGAURL <- renderUI({

       a("Click Here to Authorise Your Google Analytics Access",
          href=ShinyGetTokenURL(securityCode)
          )
       })

  AccessToken <- reactive({
      validate(
        need(AuthCode(), "Authenticate To See")
      )

      access_token <- ShinyGetToken(code = AuthCode())

      token <- access_token$access_token

    })
}
}
}
\seealso{
Shortcut using \code{\link{doAuthMacro}}.

Other authentication functions: \code{\link{authReturnCode}};
  \code{\link{createCode}}; \code{\link{doAuthMacro}};
  \code{\link{shinygaGetToken}}
}

