% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shinyga_fetchdata.R
\name{shinygaGetUsers}
\alias{shinygaGetUsers}
\title{Get GA Users}
\usage{
shinygaGetUsers(token, accountId, start = 1, max = 1000)
}
\arguments{
\item{token}{Token passed from shinygaGetToken()}

\item{accountId}{AccountId of webproperty}

\item{max}{Maximum number to fetch}
}
\value{
If token exists, a dataframe of user permissions for the account.
}
\description{
Gets the User permissions for each account
}
\examples{
\dontrun{
aw <- shinygaGetUsers(123456)
}
}
\seealso{
Other fetch data functions: \code{\link{MEgetData}};
  \code{\link{getAndMergeGAAccounts}};
  \code{\link{processManagementData}};
  \code{\link{rollupGA}}; \code{\link{shinygaGetAccounts}};
  \code{\link{shinygaGetAdWords}};
  \code{\link{shinygaGetCustomDataSources}};
  \code{\link{shinygaGetCustomDimensions}};
  \code{\link{shinygaGetCustomMetrics}};
  \code{\link{shinygaGetFilters}};
  \code{\link{shinygaGetGoals}};
  \code{\link{shinygaGetProfiles}};
  \code{\link{shinygaGetSegments}};
  \code{\link{shinygaGetWebProperties}}
}

