% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyAnalytics.R
\name{useGoogleAnalytics}
\alias{useGoogleAnalytics}
\title{Set up a shiny app to use Google Analytics}
\usage{
useGoogleAnalytics(...)
}
\description{
This function must be called from the app's UI to
enable tracking
}

