% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shinyga_messages.R
\name{addMessageData}
\alias{addMessageData}
\title{addMessageData - creates another Message object}
\usage{
addMessageData(messageData, addText, addIcon = "users",
  addStatus = "success")
}
\arguments{
\item{messageData}{Pass in the existing messageObject created by initMessageData.}

\item{addText}{The message to add.}

\item{addIcon}{Icon taken from shinydashboard.  See Details.}

\item{addStatus}{Color of message, see Details.}
}
\value{
Nothing.
}
\description{
Use this after initMessageData to add a message when stuff happens
Icons from shinydashboard functions
}
\examples{
\dontrun{
shinyServer(function(input, output, session) {

    messageData <- initMessageData()

    addMessageData(messageData, "A new message!")

}
}
}
\seealso{
\url{http://rstudio.github.io/shinydashboard/appearance.html#icons}

Other message functions: \code{\link{initMessageData}}
}

