% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/shinyga_auth.R
\name{authReturnCode}
\alias{authReturnCode}
\title{Returns the authentication parameter "code" in redirected URLs}
\usage{
authReturnCode(session, securityCode)
}
\arguments{
\item{session}{A session object within a shinyServer function.}

\item{securityCode}{A randomly generate security code passed previously in shinygaGetTokenURL.}
}
\value{
The authentication code from the redirect URL parameter.
}
\description{
Returns the authentication parameter "code" in redirected URLs
}
\examples{
\dontrun{
securityCode <- paste0(sample(c(1:9, LETTERS, letters), 20, replace = T), collapse='')
shinyServer(function(input, output, session)){

  AuthCode <- reactive({

      authReturnCode(session, securityCode)

  })

  output$AuthGAURL <- renderUI({

       a("Click Here to Authorise Your Google Analytics Access",
          href=ShinyGetTokenURL(securityCode)
          )
       })

  AccessToken <- reactive({
      validate(
        need(AuthCode(), "Authenticate To See")
      )

      access_token <- ShinyGetToken(code = AuthCode())

      token <- access_token$access_token

    })
}
}
}
\seealso{
Other authentication functions: \code{\link{shinygaGetTokenURL}};
  \code{\link{shinygaGetToken}}
}

